pipelines

+++ scripts/pkg_hyperfine.bash
@@ -13,10 +13,10 @@ rm -rf $TMP_DIR/extracted
 rm -rf $TMP_DIR/binaries
 rm -rf $ROOT_DIR/binaries/$PACKAGE-$VERSION
 
-# if [ $(gh-release-exists alshdavid/xpkg "${PACKAGE}-${VERSION}") = "true" ]; then
-#   echo Already Up To Date
-#   exit 0
-# fi
+if [ $(gh-release-exists alshdavid/xpkg "${PACKAGE}-${VERSION}") = "true" ]; then
+  echo Already Up To Date
+  exit 0
+fi
 
 mkdir -p $TMP_DIR/download          
 wget -q -O $TMP_DIR/download/$PACKAGE-$VERSION.tar.gz https://github.com/sharkdp/hyperfine/releases/download/v$VERSION/hyperfine-v$VERSION-x86_64-unknown-linux-musl.tar.gz
